init mac support

+++ test/test-bb.h
@@ -165,7 +165,7 @@ private:
 
 class ArrayOutput : public BuildingBlock<ArrayOutput> {
 public:
-    GeneratorParam<std::size_t> len{"len", 5};
+    GeneratorParam<std::uint64_t> len{"len", 5};
 
     Input<Halide::Func> input{"input", Int(32), 2};
     Output<Halide::Func[]> array_output{"array_output", Int(32), 2};
